add parsing of longer sysex messages
git-svn-id: https://mididuino.googlecode.com/svn/trunk@379 6f8c04f6-b8b6-11de-9846-1b058f2a8eb1

+++ hardware/libraries/MD/Elektron.cpp
@@ -1,10 +1,9 @@
 #include "Elektron.hh"
 
-
-uint8_t data_to_sysex_elektron(uint8_t *data, uint8_t *sysex, uint8_t len) {
-  uint8_t retlen = 0;
-  uint8_t cnt;
-  uint8_t cnt7 = 0;
+uint16_t data_to_sysex_elektron(uint8_t *data, uint8_t *sysex, uint16_t len) {
+  uint16_t retlen = 0;
+  uint16_t cnt;
+  uint16_t cnt7 = 0;
 
   sysex[0] = 0;
   for (cnt = 0; cnt < len; cnt++) {
@@ -23,10 +22,10 @@ uint8_t data_to_sysex_elektron(uint8_t *data, uint8_t *sysex, uint8_t len) {
   return retlen + cnt7 + (cnt7 != 0 ? 1 : 0);
 }
 
-uint8_t sysex_to_data_elektron(uint8_t *sysex, uint8_t *data, uint8_t len) {
-  uint8_t cnt;
-  uint8_t cnt2 = 0;
-  uint8_t bits = 0;
+uint16_t sysex_to_data_elektron(uint8_t *sysex, uint8_t *data, uint16_t len) {
+  uint16_t cnt;
+  uint16_t cnt2 = 0;
+  uint16_t bits = 0;
   for (cnt = 0; cnt < len; cnt++) {
     if ((cnt % 8) == 0) {
       bits = sysex[cnt];
